[Responsive] Cleanup by removing deprecated implementation.
